9 location service (#10)
#9 location service

+++ rac-car-service/src/main/java/pl/mg/rac/car/application/dto/response/AddCarResponse.java
@@ -2,9 +2,11 @@ package pl.mg.rac.car.application.dto.response;
 
 import pl.mg.rac.commons.value.Location;
 
+import java.math.BigDecimal;
+
 public record AddCarResponse(
         String vin,
         Location location,
-        double mileage
+        BigDecimal mileage
 ) {
 }
