+++ rac-commons/src/main/java/pl/mg/rac/commons/event/car/payload/CarReturnFailedAlreadyReturnedPayload.java
@@ -3,6 +3,8 @@ package pl.mg.rac.commons.event.car.payload;
 import pl.mg.rac.commons.event.RacEventPayload;
 import pl.mg.rac.commons.value.Location;
 
+import java.math.BigDecimal;
+
 public record CarReturnFailedAlreadyReturnedPayload(String vin, String rentalId, Location location,
-                                                    double distanceTraveled) implements RacEventPayload {
+                                                    BigDecimal distanceTraveled) implements RacEventPayload {
 }
