+++ rac-commons/src/main/java/pl/mg/rac/commons/event/rentreturn/payload/ReturnRequestUserPayload.java
@@ -2,12 +2,13 @@ package pl.mg.rac.commons.event.rentreturn.payload;
 
 import pl.mg.rac.commons.event.RacEventPayload;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record ReturnRequestUserPayload(String rentId,
                                        String username,
                                        String vin,
-                                       double distanceTraveled,
+                                       BigDecimal distanceTraveled,
                                        Instant rentStartDate
 ) implements RacEventPayload {
 }
